Remove `create-secrets` command (and introduce protocol version 1.1.0)
Having `create-secrets` and `create-configurations` separate was highly
confusing, especially because most of the times, packages would just
copy secrets into the configuration anyway (because that's where they
usually need to be for nodes to use them).

To simplify we decided to remove `create-secrets` and generate secrets
within `create-configurations` if they aren't provided by the user.

As part of this change I also moved the creation of the `configs`
directory into the `FileConfigurator`. This allows other implementations
of a `Configurator` which save configs somewhere else (e.g. a `vault`
configurator could save configs directly in vault).
