+++ pkg/fileutil/fileutil.go
@@ -4,6 +4,9 @@ import (
 	"fmt"
 	"io"
 	"os"
+	"path/filepath"
+
+	homedir "github.com/mitchellh/go-homedir"
 )
 
 // Copy a file only if it doesn't exist yet
@@ -38,4 +41,30 @@ func CopyFile(src, dst string) error {
 		return err
 	}
 	return out.Close()
+}
+
+func MakeDirectory(baseDir string, subDirs ...string) (string, error) {
+	expandedBaseDir, err := homedir.Expand(baseDir)
+	if err != nil {
+		return "", err
+	}
+
+	subDirs = append([]string{expandedBaseDir}, subDirs...)
+
+	path := filepath.Join(subDirs...)
+
+	// Create directory structure if it doesn't exist
+	err = os.MkdirAll(path, os.ModePerm)
+	return path, err
+}
+
+func FileExists(name string) (bool, error) {
+	if _, err := os.Stat(name); err != nil {
+		if os.IsNotExist(err) {
+			return false, nil
+		}
+
+		return false, err
+	}
+	return true, nil
 }
